Reference file coordination and allow output of all due items
A reference file can contain similar but potentially different or corrected text - this detects discrepancies and updates from the reference file if approved by the user. Also all current due phrases can be outputted to a text file, to be viewed as a single list. All are marked as complete in the process. This "random access" approach might actually be a better approach for learning in a single session than a forward iterating list.